pyproject.toml
[tool.poetry]
name = "shushupe"
version = "0.1.0"
description = "Shushupe is a free and open source blogging platform."
homepage = "https://gahd.net/shushupe/"
authors = ["Gustavo Adolfo Huarcaya Delgado <diavolo@gahd.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
mistune = "^3.1.4"
pygments = "^2.19.2"
python-dotenv = "^1.1.1"
gunicorn = "^23.0.0"
sentry-sdk = "^2.37.0"
psycopg = {extras = ["binary"], version = "^3.2.9"}
nanoid = "^2.0.0"
html2text = "^2025.4.15"
beautifulsoup4 = "^4.13.5"
pillow = "^11.3.0"
django-imagekit = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pip-tools = "^7.5.0"
django-debug-toolbar = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
